refactor: merge test_mrope into test_position_embedding.py
